server/requirements.txt
Django==2.0.6
django-appconf==1.0.2
django-cors-headers==2.4.0
django-imagekit==4.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
pep8==1.7.1
pilkit==2.0
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
six==1.11.0
